containsKey() method implemented into TreeMap
This allows me to search through the underlying BinaryTree for a
key. I stopped last night because this exact code was not
returning true for present keys. I logged on today and it ran
first time! Very nice... I plan to do get() method next to return
a specifc keys value