Adjusted Entity/TreeMap classes to take only keys
The issue I was having was that when I tried to append to the
list of an Entity I would get an overload error. To overcome this
my plan is to allow my TreeMap to put Entities with just keys and
allow for values to be added later.

This way I can create the list first, then add it as a value for
a key